Add `GetPlayCallStats` endpoint and repository support
- Extend `TeamController` with GET /api/teams/{id}/play-call-stats to retrieve play call statistics.
- Add `PlayCallStatsDto`, `PlayCallRushDto`, and `PlayCallPassDto` for detailed response structure.
- Update `ITeamRepository` with `GetPlayCallStatsAsync` method and implement it in the repository.
- Replace `long` with `decimal` for `Sequence` in `Play`, `GameEvent`, and related entities.
- Remove unused indices and properties like `Down`, `PossessionTeamId`.
- Add database migration `RemoveDownAndYfd` to apply schema changes.

+++ PatriotIndex.Domain/Entities/GameEvent.cs
@@ -15,7 +15,7 @@ public class GameEvent : IGameEvent
     [MaxLength(50)]
     public string EventType { get; set; } = string.Empty;
 
-    public long Sequence { get; set; }
+    public decimal Sequence { get; set; }
 
     [Required]
     [MaxLength(10)]
